feat(tests): refactor integration tests for Kraken API and add PostgreSQL integration check
